requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
celery>=5.2.0
redis>=4.5.0
python-decouple>=3.8
djangorestframework-simplejwt>=5.2.0
Pillow>=10.0.0
